Update and minor bug fix
Add more details to the *disp* method of the **JSSPInstance** class and add functionality to return information as a char array, so that it can be used in other methods (e.g. to nest calls to provide detailed information).

Also, update the *getRuleAction* method of the **ruleBasedSelectionHH** class to support the *instance* argument and to fix an error when using the traditional HH model.

Closes #46